Errors Fixed

+++ Models/OdcanitUserData.cs
@@ -6,12 +6,12 @@ namespace Odmon.Worker.Models
     {
         public string? PageName { get; set; }
         public string? FieldName { get; set; }
-        public int TikCounter { get; set; }
+        public long TikCounter { get; set; }
         public string? strData { get; set; }
         public DateTime? dateData { get; set; }
         public double? numData { get; set; }
         public string? Data { get; set; }
-        public int? RowNum { get; set; }
+        public long? RowNum { get; set; }
     }
 }
